* ext/extmk.rb (extract_makefile): nothing to be removed when no file   was deleted.
* ext/extmk.rb (extmake): restore srcdir.


git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@8205 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/extmk.rb
@@ -60,9 +60,11 @@ def extract_makefile(makefile, keep = true)
   unless oldrb == newrb
     if $extout
       newrb.each {|f| installrb.delete(f)}
-      config = CONFIG.dup
-      install_dirs(target_prefix).each {|var, val| config[var] = val}
-      FileUtils.rm_f(installrb.values.collect {|f| Config.expand(f, config)}, verbose: true)
+      unless installrb.empty?
+        config = CONFIG.dup
+        install_dirs(target_prefix).each {|var, val| config[var] = val}
+        FileUtils.rm_f(installrb.values.collect {|f| Config.expand(f, config)}, verbose: true)
+      end
     end
     return false
   end
@@ -172,6 +174,7 @@ def extmake(target)
       $extpath |= $LIBPATH
     end
   ensure
+    Config::CONFIG["srcdir"] = $top_srcdir
     $hdrdir = $top_srcdir = top_srcdir
     $topdir = topdir
     Dir.chdir dir
